Refactoring: - TCPGecko classes moved to Gecko namespace. - Splitted TCPGecko classes in multiple files. - Renamed tcpconn to TCPConnector. (Naming convention fix) - Renamed Form1 to SplatAIOForm.
