+++ SplatAIO/MemAreas.cs
@@ -1,4 +1,5 @@
 ﻿using SplatAIO;
+using SplatAIO.Gecko;
 using System;
 using System.Collections.Generic;
 using System.Text;
